test: fix cross-file mock leak in dom tests
Bun's mock.module is global and persists across files. capture.dom.test
mocked ./capture-element, which leaked into capture-element.dom.test and
made its real-captureElement assertions fail depending on file order
(flaky; passed in CI by luck of ordering).

Merge both into one file with a single consistent mock set: mock
modern-screenshot + removeHiddenElements only, and exercise the real
captureElement. Order-independent now.
